feat: add mailhog to test email
Mailhog is a SMTP service that intercept the email sending at
application, store this in memory, repply to who sent the email and
don't delivery the email to the real receiver. Then, we can access all
sent email into an URL and test if the application is working fine to
send emails without spamming the real inboxes.

Signed-off-by: Vitor Mattos <vitor@php.rio>